+++ src/python/autotransform/runner/jenkins.py
@@ -12,6 +12,7 @@
 from __future__ import annotations
 
 import json
+import os
 from typing import Any, ClassVar, Dict
 
 import requests
@@ -26,44 +27,44 @@ from autotransform.runner.base import Runner, RunnerName
 from autotransform.schema.schema import AutoTransformSchema
 
 
-class JenkinsRunner(Runner):
-    """A Runner component that uses Jenkins for remote runs.
+class JenkinsAPIRunner(Runner):
+    """A Runner component that uses Jenkins API requests for remote runs.
 
     Attributes:
-        run_job_name (str): The name of the Jenkins job for running an AutoTransform schema.
-        update_job_name (str): The name of the Jenkins job for updating a Change.
+        job_name (str): The name of the Jenkins job.
         name (ClassVar[RunnerName]): The name of the component.
     """
 
-    run_job_name: str
-    update_job_name: str
+    job_name: str
 
-    name: ClassVar[RunnerName] = RunnerName.JENKINS
+    name: ClassVar[RunnerName] = RunnerName.JENKINS_API
 
     def run(self, schema: AutoTransformSchema) -> None:
-        """Triggers a full run of a Schema locally.
+        """Triggers a full run of a Schema using a Jenkins API request.
 
         Args:
             schema (AutoTransformSchema): The schema that will be run.
         """
 
-        job_params = {"schema": schema.config.schema_name}
+        job_params = {"COMMAND": "run", "SCHEMA_NAME": schema.config.schema_name}
         if schema.config.max_submissions:
-            job_params["max_submissions"] = str(schema.config.max_submissions)
+            job_params["MAX_SUBMISSIONS"] = str(schema.config.max_submissions)
         shard_filter = [filt for filt in schema.filters if isinstance(filt, ShardFilter)]
         if shard_filter:
-            job_params["filter"] = json.dumps(shard_filter[0].bundle())
+            job_params["FILTER"] = json.dumps(shard_filter[0].bundle())
 
-        self._run_jenkins_job(self.run_job_name, job_params)
+        self._run_jenkins_job(self.job_name, job_params)
 
     def update(self, change: Change) -> None:
-        """Triggers an update of the Change.
+        """Triggers an update of the Change using a Jenkins API request.
 
         Args:
             change (Change): The Change to update.
         """
 
-        self._run_jenkins_job(self.run_job_name, {"change": json.dumps(change.bundle())})
+        self._run_jenkins_job(
+            self.job_name, {"COMMAND": "update", "CHANGE": json.dumps(change.bundle())}
+        )
 
     @staticmethod
     def _run_jenkins_job(job_name: str, params: Dict[str, Any]) -> None:
@@ -118,3 +119,59 @@ class JenkinsRunner(Runner):
         except Exception as ex:
             event_handler.handle(WarningEvent({"message": "Failed triggering the Jenkins job"}))
             event_handler.handle(WarningEvent({"message": f"Error: {str(ex)}"}))
+
+
+class JenkinsFileRunner(Runner):
+    """A Runner component that creates files to trigger Jenkins jobs
+    using https://plugins.jenkins.io/parameterized-trigger/.
+
+    Attributes:
+        name (ClassVar[RunnerName]): The name of the component.
+        num_files (ClassVar[int]): The number of files created by the runner.
+    """
+
+    name: ClassVar[RunnerName] = RunnerName.JENKINS_FILE
+    num_files: ClassVar[int] = 0
+
+    def run(self, schema: AutoTransformSchema) -> None:
+        """Triggers a full run of a Schema by creating a file with the appropriate content.
+
+        Args:
+            schema (AutoTransformSchema): The schema that will be run.
+        """
+
+        job_params = {"COMMAND": "run", "SCHEMA_NAME": schema.config.schema_name}
+        if schema.config.max_submissions:
+            job_params["MAX_SUBMISSIONS"] = str(schema.config.max_submissions)
+        shard_filter = [filt for filt in schema.filters if isinstance(filt, ShardFilter)]
+        if shard_filter:
+            job_params["FILTER"] = json.dumps(shard_filter[0].bundle())
+
+        self._create_file(job_params)
+
+    def update(self, change: Change) -> None:
+        """Triggers an update of the Change using a Jenkins API request.
+
+        Args:
+            change (Change): The Change to update.
+        """
+
+        self._create_file({"COMMAND": "update", "CHANGE": json.dumps(change.bundle())})
+
+    def _create_file(self, props: Dict[str, str]) -> None:
+        """Creates the file for the Jenkins job.
+
+        Args:
+            props (Dict[str, str]): The props to add to the job.
+        """
+
+        JenkinsFileRunner.num_files += 1
+
+        cwd = os.getcwd().replace("\\", "/")
+        file_path = f"{cwd}/autotransform/jenkins/job_{JenkinsFileRunner.num_files}.txt"
+
+        os.makedirs(os.path.dirname(file_path), exist_ok=True)
+        content = "\n".join([f"{k}={v}" for k, v in props.items()])
+        with open(file_path, "w+", encoding="UTF-8") as job_file:
+            job_file.write(content)
+            job_file.flush()