+++ CHANGELOG.md
@@ -2,10 +2,9 @@
 
 ## Release 1.0.5
 
-
 #### New Components
  - JenkinsRunner
- 
+
 #### New Features
  - Added bash scripts and dockerfile to allow for deploying using Docker