+++ CHANGELOG.md
@@ -7,7 +7,8 @@
  - atmigrate-1.0.5 to migrate Manager for 1.0.5 changes. Can not handle in/not_in comparisons for ChangeStateCondition
 
 #### New Components
- - JenkinsRunner
+ - JenkinsAPIRunner
+ - JenkinsFileRunner
  - RequestAction
  - ReviewStateCondition
  - TestStateCondition